Move fetchData to App.vue
