+++ src/App.vue
@@ -1,11 +1,27 @@
 <script lang="ts">
 import Controls from './components/Controls.vue';
+import { getUrlVars } from './lib/utils';
+import store from './store';
 
 export default {
   name: 'App',
 
   components: {
     Controls,
+  },
+
+  async mounted() {
+    const { workspace, graph: networkName } = getUrlVars();
+    if (!workspace || !networkName) {
+      throw new Error(
+        `Workspace and network must be set! workspace=${workspace} network=${networkName}`,
+      );
+    }
+
+    store.dispatch.fetchNetwork({
+      workspaceName: workspace,
+      networkName,
+    });
   },
 };
 </script>
